% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{diann}
\alias{diann}
\title{Convert DIA-NN tsv output to \code{SummarizedExperiment} object}
\usage{
diann(file, ...)
}
\arguments{
\item{file}{\code{character}}

\item{...}{additional parameters given to \code{read.table}}
}
\value{
\code{SummarizedExperiment} object
}
\description{
The function \code{diann} will create a \code{SummarizedExperiment} from a
DIA-NN tsv file. 
The function \code{diann} takes as input the path to a .tsv 
file (DIA-NN output). Additional parameters can be given to the 
\code{read.table} function from the \code{utils} package.
}
\details{
\code{intensity} is set by default to \code{"none"}, 
Heuristics are run that select the sample columns based on expected 
metadata columns. Data upload may lead to insufficient
results.

\code{type} is set by default to \code{"tsv"}.
 
The function \code{diann} is a wrapper function with pre-set arguments 
\code{intensity} and \code{type} of the function \code{maxquant}. For further
information see also the help page of \code{maxquant}.
}
\examples{
file <- "path/to/diann/object.tsv"
\donttest{diann(file = file)}

}
