% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{maxquant}
\alias{maxquant}
\title{Convert MaxQuant txt, tsv, or xlsx output to 
\code{SummarizedExperiment} object}
\usage{
maxquant(
  file,
  intensity = c("iBAQ", "LFQ", "none"),
  sheet,
  type = c("tsv", "txt", "xlsx"),
  ...
)
}
\arguments{
\item{file}{\code{character}}

\item{intensity}{\code{character}, either \code{"iBAQ"}, \code{"LFQ"}, or 
\code{"none"}}

\item{sheet}{\code{character} or \code{numeric}, the name or index of the 
sheet to read data from}

\item{type}{\code{character}, either \code{"tsv"},  \code{"txt"}, or \code{"xlsx"}}

\item{...}{additional parameters given to \code{read.xlsx} (for 
\code{type = "xlsx"}) or \code{read.table}
(for \code{type = "tsv"}/\code{type = "txt"})}
}
\value{
\code{SummarizedExperiment} object
}
\description{
The function \code{maxquant} will create a \code{SummarizedExperiment} from a
MaxQuant tsv, txt, or xlsx file.
The function \code{maxquant} takes as input the path to a .tsv, .txt, or 
.xlsx file (MaxQuant output). Additional parameters can be given to the 
\code{read.xlsx} function from the \code{openxlsx} package (e.g. 
specifying the sheet name or index by \code{sheet}) or the \code{read.table} 
function from the \code{utils} (depending on the \code{type} argument) .
}
\details{
The argument \code{intensity} will specify if the \code{iBAQ} or 
\code{LFQ} values are taken. If \code{intensity} is set to \code{"none"}, 
heuristics are run that select the sample columns based on expected 
metadata columns. In the latter case, data upload may lead to insufficient
results.

The argument \code{type} will specify if the data is loaded from 
\code{tsv}, \code{txt}, or \code{xlsx} files.
}
\examples{
file <- "path/to/maxquant/object.txt"
\donttest{maxquant(file = file, intensity = "iBAQ", type = "txt")}

}
