% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{metaboscape}
\alias{metaboscape}
\title{Convert MetaboScape xlsx output to \code{SummarizedExperiment} object}
\usage{
metaboscape(file, sheet, ...)
}
\arguments{
\item{file}{\code{character}}

\item{sheet}{\code{character} or \code{numeric}, the name or index of the 
sheet to read data from}

\item{...}{additional parameters given to \code{read.xlsx}}
}
\value{
\code{SummarizedExperiment} object
}
\description{
The function \code{metaboscape} will create a \code{SummarizedExperiment} 
from a MetaboScape xlsx file. 
The function \code{metaboscape} takes as input the path to a .xlsx file 
(MetaboScape output) and additional parameters given to the \code{read.xlsx} 
function from the \code{openxlsx} package (e.g. specifying the sheet name or 
index by \code{sheet}).
}
\details{
Heuristics are run that select the sample columns based on expected 
metadata columns. Data upload may lead to insufficient results.
}
\examples{
file <- "path/to/metaboScape/object"
\donttest{metaboscape(file = file, sheet = 1)}

}
